test(extract/fortinet): fold whitelist tests into *_test.go via export_test.go
Move the internal whitelist tests out of a same-package test file into the
external csaf_test / cvrf_test packages, exposing toCriterion and the
product-ref/version helper types through export_test.go per Go convention.
